# backend/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for user authentication, complaints, and management.",
        "version": "1.0.0",
        "title": "Sample API",
        "contact": {
            "name": "Your Name",
            "url": "http://yourwebsite.com",
            "email": "your.email@example.com"
        }
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/login": {
            "post": {
                "summary": "Logs in a user",
                "description": "Logs in a user with the provided credentials.",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username for login",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password for login",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged in",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/addadmin": {
            "post": {
                "summary": "Adds an admin user",
                "description": "Creates a new admin user with the provided details.",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username for admin",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password for admin",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin user created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error creating admin user"
                    }
                }
            }
        },
...
    }
}